requirements_qt.txt
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Bottleneck==1.4.2
certifi==2025.1.31
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.4.1
cleo==0.8.1
clevercsv==0.7.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
contourpy==1.3.1
crashtest==0.3.1
cryptography==3.4.8
cycler==0.12.1
Cython==0.29.28
dbus-python==1.2.18
decorator==4.4.2
deepdiff==5.6.0
defer==1.0.6
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.6.0
fonttools==4.56.0
future==0.18.2
gast==0.5.2
greenlet==1.1.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
jdcal==1.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpickle==2.0.0+dfsg1
keyring==23.5.0
keyrings.alt==4.1.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.38.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
MarkupSafe==2.0.1
matplotlib==3.10.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
netifaces==0.11.0
networkx==3.4.2
numba==0.55.1
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
openpyxl==3.0.9
ordered-set==4.1.0
packaging==21.3
pandas==2.2.3
paramiko==2.9.3
pastel==0.2.1
pbr==6.1.1
pexpect==4.8.0
Pillow==9.0.1
Pint==0.18
pip==25.0.1
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
psutil==6.1.1
ptyprocess==0.7.0
py==1.10.0
pybind11==2.9.1
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==1.8.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pylev==1.2.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrosetta-installer==0.1.2
pyspnego==0.11.2
pytest==6.2.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qcelemental==0.24.0
qiskit==1.0.2
qiskit-aer==0.14.1
qiskit-algorithms==0.3.0
qiskit-ibm-provider==0.11.0
regex==2021.11.10
reportlab==3.6.8
requests==2.32.3
requests_ntlm==1.3.0
rustworkx==0.16.0
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.1
setuptools==75.8.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
stevedore==5.4.0
symengine==0.13.0
sympy==1.13.3
systemd-python==234
tables==3.7.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wheel==0.45.1
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0